glassbox_outreach_new_code_to_add/api/app/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
sqlalchemy==2.0.29
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.6.1
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.1
playwright==1.43.0
gspread==5.11.1
google-auth==2.23.0
google-api-python-client==2.99.0
python-dateutil==2.9.0
tenacity==8.3.0
email-validator==2.1.1
dnspython==2.4.2
beautifulsoup4==4.12.2